requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.5.2
MarkupSafe==2.1.3
mongoengine==0.27.0
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pytesseract==0.3.10
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.2.4
twilio==8.10.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wget==3.2
yarl==1.9.3
